+++ storage/local_test.go
@@ -5,18 +5,18 @@ import (
 	"testing"
 )
 
-func TestLocalClient_GetBucketNames(t *testing.T) {
+func TestLocalClient_GetDiskNames(t *testing.T) {
 	envName := "test"
 	directory := os.TempDir()
 	c := LocalClient{EnvName: envName, Directory: directory}
-	names, _ := c.GetBucketNames()
+	names, _ := c.GetDiskNames()
 
 	if len(names) != 1 {
-		t.Error("wrong number of bucket returned")
+		t.Error("wrong number of disk returned")
 	}
 
 	if names[0] != directory {
-		t.Errorf("BucketInfo name returned: %s, BucketInfo name exptected: %s", names[0], directory)
+		t.Errorf("DiskInfo name returned: %s, DiskInfo name exptected: %s", names[0], directory)
 	}
 
 }